+++ codex-rs/cli/tests/marketplace_remove.rs
@@ -1,7 +1,7 @@
 use anyhow::Result;
 use codex_config::MarketplaceConfigUpdate;
 use codex_config::record_user_marketplace;
-use codex_core::plugins::marketplace_install_root;
+use codex_core_plugins::installed_marketplaces::marketplace_install_root;
 use predicates::str::contains;
 use std::path::Path;
 use tempfile::TempDir;
