+++ codex-rs/core-plugins/Cargo.toml
@@ -36,5 +36,6 @@ tracing = { workspace = true }
 url = { workspace = true }
 
 [dev-dependencies]
+anyhow = { workspace = true }
 pretty_assertions = { workspace = true }
 tempfile = { workspace = true }
