+++ codex-rs/core-plugins/src/lib.rs
@@ -1,8 +1,14 @@
+pub mod installed_marketplaces;
 pub mod loader;
 pub mod manifest;
 pub mod marketplace;
+pub mod marketplace_add;
+pub mod marketplace_remove;
 pub mod marketplace_upgrade;
 pub mod remote;
 pub mod remote_legacy;
 pub mod store;
 pub mod toggles;
+
+pub const OPENAI_CURATED_MARKETPLACE_NAME: &str = "openai-curated";
+pub const OPENAI_BUNDLED_MARKETPLACE_NAME: &str = "openai-bundled";
