+++ codex-rs/core-plugins/src/marketplace_add/metadata.rs
@@ -1,10 +1,10 @@
 use super::MarketplaceAddError;
 use super::source::MarketplaceSource;
-use crate::plugins::installed_marketplaces::resolve_configured_marketplace_root;
+use crate::installed_marketplaces::resolve_configured_marketplace_root;
+use crate::marketplace::validate_marketplace_root;
 use codex_config::CONFIG_TOML_FILE;
 use codex_config::MarketplaceConfigUpdate;
 use codex_config::record_user_marketplace;
-use codex_core_plugins::marketplace::validate_marketplace_root;
 use std::fs;
 use std::io::ErrorKind;
 use std::path::Path;
