+++ codex-rs/core/src/plugins/discoverable.rs
@@ -2,12 +2,12 @@ use anyhow::Context;
 use std::collections::HashSet;
 use tracing::warn;
 
-use super::OPENAI_BUNDLED_MARKETPLACE_NAME;
-use super::OPENAI_CURATED_MARKETPLACE_NAME;
 use super::PluginCapabilitySummary;
 use super::PluginsManager;
 use crate::config::Config;
 use codex_config::types::ToolSuggestDiscoverableType;
+use codex_core_plugins::OPENAI_BUNDLED_MARKETPLACE_NAME;
+use codex_core_plugins::OPENAI_CURATED_MARKETPLACE_NAME;
 use codex_features::Feature;
 use codex_tools::DiscoverablePluginInfo;
