+++ codex-rs/core/src/plugins/mod.rs
@@ -4,8 +4,6 @@ mod discoverable;
 mod injection;
 mod installed_marketplaces;
 mod manager;
-mod marketplace_add;
-mod marketplace_remove;
 mod mentions;
 mod render;
 mod startup_sync;
@@ -27,13 +25,9 @@ pub type PluginLoadOutcome = codex_plugin::PluginLoadOutcome<McpServerConfig>;
 
 pub(crate) use discoverable::list_tool_suggest_discoverable_plugins;
 pub(crate) use injection::build_plugin_injections;
-pub use installed_marketplaces::INSTALLED_MARKETPLACES_DIR;
-pub use installed_marketplaces::marketplace_install_root;
 pub use manager::ConfiguredMarketplace;
 pub use manager::ConfiguredMarketplaceListOutcome;
 pub use manager::ConfiguredMarketplacePlugin;
-pub use manager::OPENAI_BUNDLED_MARKETPLACE_NAME;
-pub use manager::OPENAI_CURATED_MARKETPLACE_NAME;
 pub use manager::PluginDetail;
 pub use manager::PluginDetailsUnavailableReason;
 pub use manager::PluginInstallError;
@@ -45,15 +39,6 @@ pub use manager::PluginRemoteSyncError;
 pub use manager::PluginUninstallError;
 pub use manager::PluginsManager;
 pub use manager::RemotePluginSyncResult;
-pub use marketplace_add::MarketplaceAddError;
-pub use marketplace_add::MarketplaceAddOutcome;
-pub use marketplace_add::MarketplaceAddRequest;
-pub use marketplace_add::add_marketplace;
-pub use marketplace_add::is_local_marketplace_source;
-pub use marketplace_remove::MarketplaceRemoveError;
-pub use marketplace_remove::MarketplaceRemoveOutcome;
-pub use marketplace_remove::MarketplaceRemoveRequest;
-pub use marketplace_remove::remove_marketplace;
 pub(crate) use render::render_explicit_plugin_instructions;
 pub(crate) use startup_sync::curated_plugins_repo_path;
 pub(crate) use startup_sync::read_curated_plugins_sha;
