+++ codex-rs/tui/src/chatwidget/tests.rs
@@ -19,7 +19,6 @@ pub(super) use crate::legacy_core::config::Config;
 pub(super) use crate::legacy_core::config::ConfigBuilder;
 pub(super) use crate::legacy_core::config::Constrained;
 pub(super) use crate::legacy_core::config::ConstraintError;
-pub(super) use crate::legacy_core::plugins::OPENAI_CURATED_MARKETPLACE_NAME;
 pub(super) use crate::model_catalog::ModelCatalog;
 pub(super) use crate::test_backend::VT100Backend;
 pub(super) use crate::test_support::PathBufExt;
@@ -106,6 +105,7 @@ pub(super) use codex_config::types::ApprovalsReviewer;
 pub(super) use codex_config::types::Notifications;
 #[cfg(target_os = "windows")]
 pub(super) use codex_config::types::WindowsSandboxModeToml;
+pub(super) use codex_core_plugins::OPENAI_CURATED_MARKETPLACE_NAME;
 pub(super) use codex_core_skills::model::SkillMetadata;
 pub(super) use codex_features::FEATURES;
 pub(super) use codex_features::Feature;
